+++ Database/models/CreateAutonomousContainerDatabaseDetails.cs
@@ -181,5 +181,11 @@ namespace Oci.DatabaseService.Models
         /// </value>
         [JsonProperty(PropertyName = "vaultId")]
         public string VaultId { get; set; }
+
+        /// <value>
+        /// The [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the key store.
+        /// </value>
+        [JsonProperty(PropertyName = "keyStoreId")]
+        public string KeyStoreId { get; set; }
     }
 }