+++ Loggingsearch/models/FieldInfo.cs
@@ -34,7 +34,7 @@ namespace Oci.LoggingsearchService.Models
         /// <value>
         /// Field type -
         /// * `STRING`: A sequence of characters.
-        /// * `NUMBER`: Numeric type which can be integer or floating point.
+        /// * `NUMBER`: Numeric type which can be an integer or floating point.
         /// * `BOOLEAN`: Either true or false.
         /// * `ARRAY`: An ordered collection of values.
         /// 
@@ -54,7 +54,7 @@ namespace Oci.LoggingsearchService.Models
         /// <value>
         /// Field type -
         /// * `STRING`: A sequence of characters.
-        /// * `NUMBER`: Numeric type which can be integer or floating point.
+        /// * `NUMBER`: Numeric type which can be an integer or floating point.
         /// * `BOOLEAN`: Either true or false.
         /// * `ARRAY`: An ordered collection of values.
         ///